requirements.txt
Cython
numpy
matplotlib
requests
pycocotools-windows
pillow
opencv-python
opencv-contrib-python
fastdtw
scipy